% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\docType{data}
\name{columns}
\alias{columns}
\title{Column metadata}
\format{Data frame with three columns:
\describe{
\item{feed}{Feed name}
\item{column}{Column name}
\item{description}{Textual description}
}}
\usage{
columns
}
\description{
Data frame storing descriptions of fields/columns returned by various feeds.
}
\keyword{datasets}
