% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft.R
\name{fse_aircraft}
\alias{fse_aircraft}
\alias{fse_aircraft_by_makemodel}
\alias{fse_aircraft_by_key}
\alias{fse_aircraft_by_owner_name}
\alias{fse_icao_aircraft}
\alias{fse_aircraft_for_sale}
\title{Query FSE aircraft}
\usage{
fse_aircraft_by_makemodel(makemodel, ...)

fse_aircraft_by_key(accesskey, ...)

fse_aircraft_by_owner_name(ownername, ...)

fse_icao_aircraft(icao, ...)

fse_aircraft_for_sale(...)
}
\arguments{
\item{makemodel}{character, aircraft model name, see Details}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}

\item{accesskey}{FSE read access key}

\item{ownername}{character, FSE account name}

\item{icao}{airport ICAO code}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

All these function return data.frame/XML with the following columns/tags:

\itemize{
\item \code{SerialNumber} -- Aircraft serial number
\item \code{MakeModel} -- Aircraft model name
\item \code{Registration} -- Registration number
\item \code{Owner}
\item \code{Location}
\item \code{LocationName}
\item \code{Home}
\item \code{SalePrice}
\item \code{SellbackPrice}
\item \code{Equipment}
\item \code{RentalDry}
\item \code{RentalWet}
\item \code{Bonus}
\item \code{RentalTime}
\item \code{RentedBy}
\item \code{PctFuel}
\item \code{NeedsRepair}
\item \code{AirframeTime}
\item \code{EngineTime}
\item \code{TimeLast100hr}
\item \code{LeasedFrom}
\item \code{MonthlyFee}
\item \code{FeeOwed}
}
}
\description{
Functions for querying FSE aircraft.

\itemize{
\item \code{fse_aircraft_by_makemodel} queries aircraft by model name
}

\itemize{
\item \code{fse_aircraft_by_key} queries for aircraft associated to an
FSE read access key
}

\itemize{
\item \code{fse_aircraft_by_owner_name} query aircraft by owner name
}

\itemize{
\item \code{fse_icao_aircraft} -- List of aircraft at ICAO
}

\itemize{
\item \code{\link[=fse_aircraft_for_sale]{fse_aircraft_for_sale()}} -- List aircraft currently for sale
}
}
\details{
For \code{\link[=fse_aircraft_by_makemodel]{fse_aircraft_by_makemodel()}}, to get a list of valid model
names that can be submitted to \code{makemodel} fetch aircraft configs with
\code{\link{fse_aicraft_configs}} or go to FSE system WWW page, "Aircraft"
menu, "Purchase aircraft" item and check the dropdown list after selecting
"By model" radio button.
}
\examples{
\dontrun{
r <- fse_aircraft_by_makemodel("Cessna 172 Skyhawk")
}
\dontrun{
r <- fse_icao_aircraft("epbc")
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo}},
\code{\link{fse_group_members}()},
\code{\link{fse_jobs}},
\code{\link{fse_log}},
\code{\link{fse_payments}},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
