% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_aliases.R
\name{fse_aircraft_aliases}
\alias{fse_aircraft_aliases}
\title{Aircraft Aliases}
\usage{
fse_aircraft_aliases(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{link{fse_api}}).
}
\description{
Aliases of FSE aircraft.
}
\seealso{
Other fse_feeds: \code{\link{fse_aircraft_configs}},
  \code{\link{fse_aircraft_for_sale}},
  \code{\link{fse_aircraft_status_by_registration}}
}

