% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_by_makemodel.R
\name{fse_aircraft_by_makemodel}
\alias{fse_aircraft_by_makemodel}
\title{Aircraft by MakeModel}
\usage{
fse_aircraft_by_makemodel(makemodel, ...)
}
\arguments{
\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{\link{fse_api}}).
}
\description{
Aircraft by MakeModel
}
\seealso{
Other fse_feeds: \code{\link{fse_aircraft_aliases}},
  \code{\link{fse_aircraft_by_owner_name}},
  \code{\link{fse_aircraft_configs}},
  \code{\link{fse_aircraft_for_sale}},
  \code{\link{fse_aircraft_status_by_registration}},
  \code{\link{fse_icao_jobs_to}}
}

