% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_configs.R
\docType{data}
\name{fse_aircraft_configs}
\alias{fse_aircraft_configs}
\alias{aircraft_configs}
\title{Aircraft configuration data}
\format{
Object \code{aircraft_configs} is a data frame returned by
\code{\link[=fse_aircraft_configs]{fse_aircraft_configs()}} and has the following columns:

\itemize{
\item \code{MakeModel} -- Aircraft model name
\item \code{Crew} -- Number of crew members
\item \code{Seats} -- Number of PAX seats
\item \code{CruiseSpeed} -- Average cruise speed
\item \code{GPH} -- Fuel consumption in Gallons Per Hour
\item \code{FuelType} -- Fuel type
\item \code{MTOW} -- Maximal TakeOff Weight
\item \code{EmptyWeight} -- Empty weight
\item \code{Price} -- System aircraft price
\item \code{Ext1} -- Fuel tank capacity: external 1
\item \code{LTip} --
\item \code{LAux} --
\item \code{LMain} --
\item \code{Center1} --
\item \code{Center2} --
\item \code{Center3} --
\item \code{RMain} --
\item \code{RAux} --
\item \code{RTip} --
\item \code{RExt2} -- Fuel tank capacity: right external 2
\item \code{Engines} -- Number of engines
\item \code{EnginePrice} -- Engine price (maintenance)
\item \code{ModelId} -- Aircraft model ID
}
}
\usage{
fse_aircraft_configs(...)

aircraft_configs
}
\arguments{
\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

Function \code{\link[=fse_aircraft_configs]{fse_aircraft_configs()}} return a data frame or XML object
with the columns/tags listed in the Format section. Object
\link{aircraft_configs} is simply a data frame returned by that function on
2020-04-03.
}
\description{
Data on all types of aircraft available in FSE. The function will fetch it
from the server, but it does not change too frequently, hence a built-in
version too.
}
\examples{
\dontrun{
r <- fse_aircraft_configs()
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo_summary}()},
\code{\link{fse_fbo}},
\code{\link{fse_group_members}()},
\code{\link{fse_jobs}},
\code{\link{fse_log}},
\code{\link{fse_payments}},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
\keyword{datasets}
