% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_status_by_registration.R
\name{fse_aircraft_status_by_registration}
\alias{fse_aircraft_status_by_registration}
\title{Aircraft Status by Registration}
\usage{
fse_aircraft_status_by_registration(aircraftreg, ...)
}
\arguments{
\item{aircraftreg}{aircraft registration number}

\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
\code{xml_document} that can be processed with \pkg{xml2}
}
\description{
Lookup the status of a particular aircraft.
}
\examples{
\dontrun{
r <- fse_aircraft_status_by_registration(483514)
xml2::xml_structure(r)
}
}

