% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_fbo.R
\name{fse_fbo}
\alias{fse_fbo}
\alias{fse_fbo_by_key}
\alias{fse_fbo_for_sale}
\alias{fse_icao_fbos}
\title{Query FSE FBOs}
\usage{
fse_fbo_by_key(accesskey, ...)

fse_fbo_for_sale(...)

fse_icao_fbos(icao, ...)
}
\arguments{
\item{accesskey}{FSE read access key}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

The object with the following columns/tags:

\itemize{
\item \code{FboId} -- FBO id
\item \code{Status} -- Status of the FBO: "Active" or "Closed"
\item \code{Airport} -- Name of the airport
\item \code{Name} -- Name of the FBO
\item \code{Owner} -- Owner
\item \code{Icao} -- ICAO code of the airport
\item \code{Location} -- Name of the location
\item \code{Lots} -- Number of lots
\item \code{RepairShop} -- Is there a repair shop: "Yes" or "No"
\item \code{Gates} -- Number of gates or "No Passenger Terminal"
\item \code{GatesRented} -- Number of gates rented or "No Passenger Terminal"
\item \code{Fuel100LL} -- Available 100LL in kilograms
\item \code{FuelJetA} -- Available JetA in kilograms
\item \code{BuildingMaterials} -- Available building materials in kilograms
\item \code{Supplies} -- Available supplies in kilograms
\item \code{SuppliesPerDay} -- Supplies consumption per day
\item \code{SuppliedDays} -- For how many days the FBO will be supplied
\item \code{SellPrice} -- ?
\item \code{Fuel100LLGal} -- Available 100L fuel in galons
\item \code{FuelJetAGal} -- Available JetA fuel in galons
\item \code{Price100LLGal} -- Price of 100L fuel per galon
\item \code{PriceJetAGal} -- Price of JetA fuel per galon
}
}
\description{
These function query FSE FBOs:

\itemize{
\item \code{\link[=fse_fbo_by_key]{fse_fbo_by_key()}} -- list of FBOs belonging to the provided read access key.
}

\itemize{
\item \code{\link[=fse_fbo_for_sale]{fse_fbo_for_sale()}} - list of FBOs for sale
}

\itemize{
\item \code{\link[=fse_icao_fbos]{fse_icao_fbos()}} - list FBOs present at an airport
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_group_members}()},
\code{\link{fse_jobs}},
\code{\link{fse_log}},
\code{\link{fse_payments}},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
