% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_fbo_summary.R
\name{fse_fbo_summary}
\alias{fse_fbo_summary}
\title{Monthly summary statistics for an FBO}
\usage{
fse_fbo_summary(icao, month, year, accesskey, ...)
}
\arguments{
\item{icao}{Airport ICAO code}

\item{year, month}{year and month as integers}

\item{accesskey}{FSE access key}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

Columns to be provided.
}
\description{
Monthly summary statistics for an FBO
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo}},
\code{\link{fse_group_members}()},
\code{\link{fse_jobs}},
\code{\link{fse_log}},
\code{\link{fse_payments}},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
