% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_icao_aircraft.R
\name{fse_icao_aircraft}
\alias{fse_icao_aircraft}
\title{ICAO listing of aircraft}
\usage{
fse_icao_aircraft(icao, ...)
}
\arguments{
\item{icao}{airport ICAO code}

\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{\link{fse_api}}). In either case the result will
contain the following columns/tags:

\Sexpr[results=verbatim,stage=render,echo=FALSE]{try(fse:::column_descriptions('icao_fbos'))}
}
\description{
List aircraft at location.
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_by_key}()},
\code{\link{fse_aircraft_by_makemodel}()},
\code{\link{fse_aircraft_by_owner_name}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo_by_key}()},
\code{\link{fse_fbo_for_sale}()},
\code{\link{fse_group_members}()},
\code{\link{fse_icao_fbos}()},
\code{\link{fse_icao_jobs_from}()},
\code{\link{fse_icao_jobs_to}()},
\code{\link{fse_log_by_reg_from_id}()},
\code{\link{fse_log_by_serial_from_id}()},
\code{\link{fse_payments_by_monthyear}()},
\code{\link{fse_payments_from_id}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
