% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_jobs.R
\name{fse_jobs}
\alias{fse_jobs}
\alias{fse_icao_jobs_from}
\alias{fse_icao_jobs_to}
\title{Assignments to and from airports}
\usage{
fse_icao_jobs_from(icaos, ...)

fse_icao_jobs_to(icaos, ...)
}
\arguments{
\item{icaos}{airport ICAO codes (can be more than one, as a character vector)}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

The object with the following columns/tags:

\itemize{
\item \code{Id} -- Assignment Id
\item \code{Location} -- ICAO where the payload is currently at
\item \code{ToIcao} -- ICAO of destination airport
\item \code{FromIcao} -- ICAO of origin airport
\item \code{Amount} -- Payload amount
\item \code{UnitType} -- The unit of \code{Amount}
\item \code{Commodity} -- Payload description
\item \code{Pay} -- Assignment pay in v$
\item \code{Expires} -- Time in which the assignment expires
\item \code{ExpireDateTime} -- Date and time at which the assignment expires
\item \code{Express} -- Is it an express assignment: "False", "True"
\item \code{PtAssignment} -- Is it an assignment of a PAX terminal: "False", "True"
\item \code{Type} -- Type of assignment: "Trip-Only", "All-In"
\item \code{AircraftId} -- For "All-In" assignments the serial number of the aircraft assigned to the job
}
}
\description{
\itemize{
\item \code{\link[=fse_icao_jobs_from]{fse_icao_jobs_from()}} -- list jobs from airports with
provided ICAO codes
}

\itemize{
\item \code{\link[=fse_icao_jobs_to]{fse_icao_jobs_to()}} -- list jobs to airports with provided ICAO codes
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo}},
\code{\link{fse_group_members}()},
\code{\link{fse_log}},
\code{\link{fse_payments_by_monthyear}()},
\code{\link{fse_payments_from_id}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
