% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_log.R
\name{fse_log}
\alias{fse_log}
\alias{fse_log_by_serial_from_id}
\alias{fse_log_by_key_from_id}
\alias{fse_log_by_reg_from_id_group}
\alias{fse_log_by_reg_from_id}
\alias{fse_log_by_serial_monthyear}
\alias{fse_log_by_reg_monthyear}
\title{Query FSE flight logs}
\usage{
fse_log_by_serial_from_id(serialnumber, fromid, ...)

fse_log_by_key_from_id(accesskey, fromid, ...)

fse_log_by_reg_from_id_group(accesskey, fromid, ...)

fse_log_by_reg_from_id(aircraftreg, fromid, ...)

fse_log_by_serial_monthyear(serialnumber, year, month, ...)

fse_log_by_reg_monthyear(aircraftreg, year, month, ...)
}
\arguments{
\item{serialnumber}{aircraft serial number}

\item{fromid}{log entry id}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}

\item{accesskey}{group's read access key}

\item{aircraftreg}{aircraft registration number}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

All these functions return data.frame/XML with the following
columns/tags:

\itemize{
\item \code{Id} -- Log entry id
\item \code{Type} -- Entry type
\item \code{Time} -- Time as \code{yyyy/mm/dd}
\item \code{Distance} -- If \code{Type = flight} distance flown, \code{0} otherwise
\item \code{Pilot} -- Pilot name
\item \code{SerialNumber} -- Aircraft serial number
\item \code{Aircraft} -- Registration id
\item \code{MakeModel} -- Aircraft make and model
\item \code{From} -- ICAO from
\item \code{To} -- ICAO to
\item \code{TotalEngineTime} -- Total engine time as \code{hh:mm}
\item \code{FlightTime} -- Flight time as \code{hh:mm}
\item \code{GroupName} -- For group flights the group name or \code{""} otherwise
\item \code{Income}
\item \code{PilotFee}
\item \code{CrewCost}
\item \code{BookingFee}
\item \code{Bonus}
\item \code{FuelCost}
\item \code{GCF} -- ?
\item \code{RentalPrice}
\item \code{RentalType} -- ("hoobs); ("tacho")
\item \code{RentalUnits} --
\item \code{RentalCost}
}
}
\description{
Functions for querying FSE flight logs.

\itemize{
\item \code{fse_log_by_serial_from_id} queries flight log for specified
aircraft \code{serialnumber} and that have id greater than \code{formid}.
}

\itemize{
\item \code{fse_log_by_key_from_id} queries flight log for entries
for associated with \code{accesskey} with id numbers geater than \code{fromid}.
}

\itemize{
\item \code{fse_log_by_reg_from_id_group} queries flight logs by key
for from id from all group aircraft. List all log entries for associated
with all the aircraft of group with \code{accesskey} and that have id number
higher than \code{fromid}.
}

\itemize{
\item \code{fse_log_by_reg_from_id} Flight log by aircraft
registration number from Id. List all log entries for associated with the
aircraft with \code{aircraftreg} and that have id number higher than \code{fromid}.
}

\itemize{
\item \code{fse_log_by_serial_monthyear} -- Query flight logs by aircraft
serial number, year and month.
}

\itemize{
\item \code{fse_log_by_reg_monthyear} -- Query flight logs by aircraft
serial number, year and month.
}
}
\examples{
\dontrun{
r <- fse_log_by_serial_from_id(32019, 1)
}
\dontrun{
r <- fse_log_by_serial_monthyear(6930, 2016, 4)
}
\dontrun{
r <- fse_log_by_reg_monthyear("SP-4AX", 2020, 2)
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo_by_key}()},
\code{\link{fse_fbo_for_sale}()},
\code{\link{fse_group_members}()},
\code{\link{fse_icao_fbos}()},
\code{\link{fse_icao_jobs_from}()},
\code{\link{fse_icao_jobs_to}()},
\code{\link{fse_payments_by_monthyear}()},
\code{\link{fse_payments_from_id}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
