% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_statistics_by_key.R
\name{fse_statistics_by_key}
\alias{fse_statistics_by_key}
\title{Statistics by read access key}
\usage{
fse_statistics_by_key(accesskey, ...)
}
\arguments{
\item{accesskey}{FSE read access key}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

The object with the following columns/tags:

\itemize{
\item \code{Personal_balance} -- Money in the wallet
\item \code{Bank_balance} -- Money in the bank
\item \code{flights} -- Number of flights made to date
\item \code{Total_Miles} -- Total number of NMs flown
\item \code{Time_Flown} -- Total time in the air in \code{hh:mm}
}
}
\description{
User statistics such as financials, number of flights and miles flown.
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo}},
\code{\link{fse_group_members}()},
\code{\link{fse_jobs}},
\code{\link{fse_log}},
\code{\link{fse_payments}}
}
\concept{FSE feeds}
