% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icao_data.R
\docType{data}
\name{icao_data}
\alias{icao_data}
\title{FSE ICAO data}
\format{
A data frame with 23760 rows and the following 9 columns:

\itemize{
\item \code{icao} -- ICAO code
\item \code{lat} -- Latitude
\item \code{lon} -- Longitude
\item \code{type} -- Type of airport: "civil", "military", "water"
\item \code{size} -- Airport size?
\item \code{name} -- Name of the airport
\item \code{city} -- Airport city name
\item \code{state} -- Airport state (for U.S.)
\item \code{country} -- Airtport country
}
}
\usage{
icao_data
}
\description{
Data on airports available in FSE.
}
\keyword{datasets}
