% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icao_distance.R
\name{icao_distance}
\alias{icao_distance}
\alias{icao_distance.character}
\title{Great circle distance between airports}
\usage{
icao_distance(x, ...)

\method{icao_distance}{character}(x, unit = c("nm", "km", "m"), ...)
}
\arguments{
\item{x}{character vector of ICAO codes}

\item{...}{other arguments}

\item{unit}{unit of the computation}
}
\description{
Great circle distance between airports
}
