% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft.R
\name{fse_aircraft}
\alias{fse_aircraft}
\alias{fse_aircraft_by_makemodel}
\alias{fse_aircraft_by_key}
\alias{fse_aircraft_by_owner_name}
\title{Query FSE aircraft}
\usage{
fse_aircraft_by_makemodel(makemodel, ...)

fse_aircraft_by_key(accesskey, ...)

fse_aircraft_by_owner_name(ownername, ...)
}
\arguments{
\item{makemodel}{character, aircraft model name, see Details

For \code{\link[=fse_aircraft_by_makemodel]{fse_aircraft_by_makemodel()}}, to get a list of valid model names that
can be submitted to \code{makemodel} fetch aircraft configs with
\code{\link{fse_aicraft_configs}} or go to FSE system WWW page, "Aircraft"
menu, "Purchase aircraft" item and check the dropdown list after selecting
"By model" radio button.}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}

\item{accesskey}{FSE read access key}

\item{ownername}{character, FSE account name}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).
}
\description{
Functions for querying FSE aircraft.

\itemize{
\item \code{fse_airceaft_by_makemodel} queries aircraft by model name
}

\itemize{
\item \code{fse_aircraft_by_key} queries for aircraft associated to an
FSE read access key
}

\itemize{
\item \code{fse_aircraft_by_owner_name} query aircraft by owner name
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo_by_key}()},
\code{\link{fse_fbo_for_sale}()},
\code{\link{fse_group_members}()},
\code{\link{fse_icao_aircraft}()},
\code{\link{fse_icao_fbos}()},
\code{\link{fse_icao_jobs_from}()},
\code{\link{fse_icao_jobs_to}()},
\code{\link{fse_log}},
\code{\link{fse_payments_by_monthyear}()},
\code{\link{fse_payments_from_id}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
