% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_by_makemodel.R
\name{fse_aircraft_by_makemodel}
\alias{fse_aircraft_by_makemodel}
\title{Aircraft by MakeModel}
\usage{
fse_aircraft_by_makemodel(makemodel, ...)
}
\arguments{
\item{makemodel}{character, aircraft model name, see Details}

\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{\link{fse_api}}). In either case the result will
contain the following columns/tags:

\Sexpr[results=verbatim,stage=render,echo=FALSE]{try(fse:::column_descriptions(''))}
}
\description{
Aircraft by MakeModel
}
\details{
To get a list of valid model names that can be submitted to \code{makemodel}
fetch aircraft configs with \code{\link{fse_aicraft_configs}} or go to FSE
system WWW page, "Aircraft" menu, "Purchase aircraft" item and check the
dropdown list after selecting "By model" radio button.
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_by_key}()},
\code{\link{fse_aircraft_by_owner_name}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_icao_jobs_from}()},
\code{\link{fse_icao_jobs_to}()}
}
\concept{FSE feeds}
