% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_configs.R
\name{fse_aircraft_configs}
\alias{fse_aircraft_configs}
\title{Aircraft configuraion data}
\usage{
fse_aircraft_configs(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{\link{fse_api}}).
}
\description{
Data on all types of aircraft available in FSE.
}
\seealso{
Other fse_feeds: \code{\link{fse_aircraft_aliases}},
  \code{\link{fse_aircraft_by_makemodel}},
  \code{\link{fse_aircraft_by_owner_name}},
  \code{\link{fse_aircraft_for_sale}},
  \code{\link{fse_aircraft_status_by_registration}},
  \code{\link{fse_icao_jobs_to}}
}

