% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_for_sale.R
\name{fse_aircraft_for_sale}
\alias{fse_aircraft_for_sale}
\title{Aircraft for Sale}
\usage{
fse_aircraft_for_sale(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{\link{fse_api}}).

The object will have the following tags/columns:

\Sexpr[results=verbatim]{data(columns); subset(columns, feed == "aircraft_for_sale", select=c("column", "description"))}
}
\description{
Aircraft for Sale
}
\seealso{
Other fse_feeds: \code{\link{fse_aircraft_aliases}},
  \code{\link{fse_aircraft_by_makemodel}},
  \code{\link{fse_aircraft_by_owner_name}},
  \code{\link{fse_aircraft_configs}},
  \code{\link{fse_aircraft_status_by_registration}},
  \code{\link{fse_icao_jobs_from}},
  \code{\link{fse_icao_jobs_to}}
}
