% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_for_sale.R
\name{fse_aircraft_for_sale}
\alias{fse_aircraft_for_sale}
\title{Aircraft for Sale}
\usage{
fse_aircraft_for_sale(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

The return object contains the following columns/tags:

\itemize{
\item \code{SerialNumber}
\item \code{MakeModel}
\item \code{Registration}
\item \code{Owner}
\item \code{Location}
\item \code{LocationName}
\item \code{Home}
\item \code{SalePrice}
\item \code{SellbackPrice}
\item \code{Equipment}
\item \code{RentalDry}
\item \code{RentalWet}
\item \code{RentalType}
\item \code{Bonus}
\item \code{RentalTime}
\item \code{RentedBy}
\item \code{PctFuel}
\item \code{NeedsRepair}
\item \code{AirframeTime}
\item \code{EngineTime}
\item \code{TimeLast100hr}
\item \code{LeasedFrom}
\item \code{MonthlyFee}
\item \code{FeeOwed}
}
}
\description{
List of aircraft currently for sale.
}
\examples{
\dontrun{
r <- fse_aircraft_for_sale()
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo_by_key}()},
\code{\link{fse_fbo_for_sale}()},
\code{\link{fse_group_members}()},
\code{\link{fse_icao_aircraft}()},
\code{\link{fse_icao_fbos}()},
\code{\link{fse_icao_jobs_from}()},
\code{\link{fse_icao_jobs_to}()},
\code{\link{fse_log}},
\code{\link{fse_payments_by_monthyear}()},
\code{\link{fse_payments_from_id}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
