% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_status_by_registration.R
\name{fse_aircraft_status_by_registration}
\alias{fse_aircraft_status_by_registration}
\title{Aircraft Status by Registration}
\usage{
fse_aircraft_status_by_registration(aircraftreg, ...)
}
\arguments{
\item{aircraftreg}{aircraft registration number}

\item{...}{other arguments passed to \code{\link[=fse_api]{fse_api()}}}
}
\value{
Feed function returns a \code{data.frame} (default) or \code{xml_document} depending on
the value of \code{format} argument (see \code{fse_api} for details).

Data.frame/XML with columns/tags:

\itemize{
\item \code{Registration} -- Aircraft registration
\item \code{SerialNumber} -- Aircraft serial number
\item \code{Status} -- Aircraft status: ("On Ground")
\item \code{Location} -- ICAO airport code if aircraft is "On Ground"
}
}
\description{
Lookup the status of a particular aircraft. Only available as an XML.
}
\examples{
\dontrun{
r <- fse_aircraft_status_by_registration(483514)
r <- fse_aircraft_status_by_registration("SP-4AX")
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo}},
\code{\link{fse_group_members}()},
\code{\link{fse_jobs}},
\code{\link{fse_log}},
\code{\link{fse_payments_by_monthyear}()},
\code{\link{fse_payments_from_id}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
