% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_aircraft_status_by_registration.R
\name{fse_aircraft_status_by_registration}
\alias{fse_aircraft_status_by_registration}
\title{Aircraft Status by Registration}
\usage{
fse_aircraft_status_by_registration(aircraftreg, ...)
}
\arguments{
\item{aircraftreg}{aircraft registration number}

\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{\link{fse_api}}). In either case the result will
contain the following columns/tags:

\Sexpr[results=verbatim,stage=render,echo=FALSE]{try(fse:::column_descriptions(''))}

This function only returns object with content as \code{xml_document} that
can be processed with \pkg{xml2}.
}
\description{
Lookup the status of a particular aircraft. Only available as an XML.
}
\examples{
\dontrun{
r <- fse_aircraft_status_by_registration(483514)
xml2::xml_structure(r)
}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_by_key}()},
\code{\link{fse_aircraft_by_makemodel}()},
\code{\link{fse_aircraft_by_owner_name}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo_by_key}()},
\code{\link{fse_fbo_for_sale}()},
\code{\link{fse_icao_jobs_from}()},
\code{\link{fse_icao_jobs_to}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
