% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_ak.R
\name{fse_ak}
\alias{fse_ak}
\alias{print.fse_key}
\title{FSE access keys}
\usage{
fse_ak(key = NULL, verbose = getOption("fse.verbose", FALSE))

\method{print}{fse_key}(x, ...)
}
\arguments{
\item{verbose}{logical, should more verbose output be provided

Retrieved from an enviroment variable \code{FSE_ACCESS_KEY}.}
}
\description{
FSE access keys
}
