% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_api.R
\name{fse_api}
\alias{fse_api}
\alias{print.fse_api}
\title{Make FSE API request}
\usage{
fse_api(
  query = c("aircraft", "assignments", "commodities", "Facilities", "fbos",
    "flightlogs", "group", "icao", "payments", "statistics"),
  search = c("status", "configs", "aliases", "forsale", "makemodel", "ownername",
    "registration", "key", "forsale", "monthlysummary", "monthyear", "id", "members",
    "aircraft", "fbo", "fbos", "jobsto", "jobsfrom"),
  format = c("csv", "xml"),
  ...,
  content_as = "parsed"
)

\method{print}{fse_api}(x, ...)
}
\arguments{
\item{query}{argument}

\item{search}{argument}

\item{format}{either \code{"csv"} (default) or \code{"xml"}}

\item{...}{other arguments, see Details}

\item{content_as}{character, passed to \code{as} argument of \code{\link{content}}}

\item{x}{object returned by \code{\link{fse_api}}}
}
\value{
A list of class "fse_api" with the following components:
\item{content}{a data frame or \code{xml_document}, depending on the value of \code{format} argument}
\item{query}{a list with query parameters used in the request}
\item{response}{the result of the request, as returned by httr::\code{\link{GET}}}
}
\description{
Function makes a request to FSE data system.
}
\details{
Requests to FSE data feeds require at least specyfying \code{format} and
\code{query} URL arguments. Possible values of these arguments are specified
in the synopsis. Do note that not all possible combinations make sense. Other
arguments, which are feed-specific can be passed through \code{...}. See FSE
feed documentation for details.

Arguments \code{format} and \code{content_as} can be used to further
customize the output format.
}
