% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_api.R
\name{fse_api}
\alias{fse_api}
\alias{print.fse_api}
\title{Make FSE API request}
\usage{
fse_api(query = c("aircraft", "assignments", "commodities", "Facilities",
  "fbos", "flightlogs", "group", "icao", "payments", "statistics"),
  search = c("status", "configs", "aliases", "forsale", "makemodel",
  "ownername", "registration", "key", "forsale", "monthlysummary", "monthyear",
  "id", "members", "aircraft", "fbo", "fbosto", "jobsfrom"), format = c("xml",
  "csv"), ..., content_as = "parsed")

\method{print}{fse_api}(x, ...)
}
\arguments{
\item{query}{argument}

\item{search}{argument}

\item{format}{either xml or csv}

\item{...}{other query arguments}

\item{content_as}{character, see \code{as} argument to \code{\link{content}}.}

\item{x}{object returned by \code{\link{fse_api}}}
}
\description{
Make FSE API request
}

