% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse_group_members.R
\name{fse_group_members}
\alias{fse_group_members}
\title{List group members}
\usage{
fse_group_members(accesskey, ...)
}
\arguments{
\item{accesskey}{Read Access Key of a FSE group}

\item{...}{other arguments passed to \code{\link{fse_api}}}
}
\value{
Feed can return \code{data.frame} (default) or \code{xml_document} depending
on the value of \code{format} (see \code{\link{fse_api}}). In either case the result will
contain the following columns/tags:

\Sexpr[results=verbatim,stage=render,echo=FALSE]{try(fse:::column_descriptions('icao_fbos'))}
}
\description{
List members of a FSE group identified by the \code{accesskey}
}
\seealso{
Other FSE feeds: 
\code{\link{fse_aircraft_aliases}()},
\code{\link{fse_aircraft_configs}()},
\code{\link{fse_aircraft_for_sale}()},
\code{\link{fse_aircraft_status_by_registration}()},
\code{\link{fse_aircraft}},
\code{\link{fse_assignments_by_key}()},
\code{\link{fse_commodities_by_key}()},
\code{\link{fse_facilities_by_key}()},
\code{\link{fse_fbo}},
\code{\link{fse_jobs}},
\code{\link{fse_log}},
\code{\link{fse_payments_by_monthyear}()},
\code{\link{fse_payments_from_id}()},
\code{\link{fse_statistics_by_key}()}
}
\concept{FSE feeds}
